Wishlisting now fully functional

+++ server/src/services/wishlist.service.ts
@@ -2,8 +2,9 @@ import mongoose from "mongoose";
 import UserModel from "../models/user.model";
 import CourseModel from "../models/course.model";
 import appAssert from "../utils/AppAssert";
-import { NOT_FOUND } from "../constants/HttpStatusCode";
+import { BAD_REQUEST, NOT_FOUND } from "../constants/HttpStatusCode";
 import WishlistModel from "../models/wishlist.model";
+import PurchaseModel from "../models/purchase.model";
 
 export const wishlistCourse = async (
   userId: mongoose.Types.ObjectId,
@@ -15,28 +16,34 @@ export const wishlistCourse = async (
   const course = await CourseModel.findById(courseId);
   appAssert(course, NOT_FOUND, "Course not found");
 
-  const wishlishedCourse = await WishlistModel.create({
+  const isCoursePurchased = await PurchaseModel.findOne({
+    user: userId,
+    course: courseId,
+  });
+  appAssert(
+    !isCoursePurchased,
+    BAD_REQUEST,
+    "Cannot wishlist a purchased course"
+  );
+
+  const isCourseWishlisted = await WishlistModel.findOne({
     user: userId,
     course: courseId,
   });
 
-  return { wishlishedCourse };
-};
+  if (isCourseWishlisted) {
+    const unwishlistedCourse = await WishlistModel.findOneAndDelete({
+      user: userId,
+      course: courseId,
+    });
 
-export const unwishlistCourse = async (
-  userId: mongoose.Types.ObjectId,
-  courseId: string
-) => {
-  const user = await UserModel.findById(userId);
-  appAssert(user, NOT_FOUND, "User not found");
+    return { message: "Course removed from wishlist successfully" };
+  } else {
+    const wishlishedCourse = await WishlistModel.create({
+      user: userId,
+      course: courseId,
+    });
 
-  const course = await CourseModel.findById(courseId);
-  appAssert(course, NOT_FOUND, "Course not found");
-
-  const wishlishedCourse = await WishlistModel.findOneAndDelete({
-    user: userId,
-    course: courseId,
-  });
-
-  return { wishlishedCourse };
+    return { message: "Course added to wishlist successfully" };
+  }
 };
